--- fs2/js/src/main/scala/org/atnos/eff/addon/fs2/package.scala
@@ -1,14 +1,0 @@
-package org.atnos.eff.addon
-
-import _root_.fs2.{Scheduler, Strategy}
-import org.atnos.eff.ExecutorServices
-
-package object fs2 extends TaskEffect {
-  /** implement in js and jvm */
-  @deprecated("The Async effect will be removed in favor of concrete asynchronous effects, like TimedFuture.", since = "2.3.0")
-  def fromExecutorServices(es: ExecutorServices): AsyncTasks = {
-    val s: Strategy = Strategy.fromExecutionContext(es.executionContext)
-    val sc: Scheduler = Scheduler.default
-    AsyncTasks()(s, sc, es)
-  }
-}
