+++ fs2/jvm/src/test/scala/org/atnos/eff/addon/fs2/TaskEffectSpec.scala
@@ -8,8 +8,10 @@ import org.specs2._
 import org.specs2.concurrent.ExecutionEnv
 
 import scala.collection.mutable.ListBuffer
-import fs2._
 import org.atnos.eff._
+import org.atnos.eff.syntax.addon.fs2.task._
+import fs2._
+import TaskEffect._
 
 import scala.concurrent.Await
 import scala.concurrent.duration._
