--- fs2/shared/src/main/scala/org/atnos/eff/syntax/addon/fs2.scala
@@ -1,5 +1,0 @@
-package org.atnos.eff.syntax.addon
-
-import org.atnos.eff.addon.fs2.TaskEffect
-
-package object fs2 extends TaskEffect
