+++ monix/jvm/src/test/scala/org/atnos/eff/addon/monix/TaskEffectSpec.scala
@@ -12,6 +12,8 @@ import scala.collection.mutable.ListBuffer
 import monix.execution.Scheduler.Implicits.global
 import monix.eval.Task
 import org.atnos.eff._
+import org.atnos.eff.addon.monix.TaskEffect._
+import org.atnos.eff.syntax.addon.monix.task._
 
 import scala.concurrent.Await
 import scala.concurrent.duration._
