+++ monix/shared/src/main/scala/org/atnos/eff/addon/monix/TaskEffect.scala
@@ -139,8 +139,6 @@ trait TaskInterpretation extends TaskTypes {
       cache.get[A](key).fold(task.map { r => cache.put(key, r); r })(Task.now)
     }
 
-  implicit final def toTaskOps[R, A](e: Eff[R, A]): TaskOps[R, A] = new TaskOps[R, A](e)
-
   /**
     * Memoize tasks using a cache
     *
@@ -172,24 +170,6 @@ trait TaskInterpretation extends TaskTypes {
     })
   }
 
-}
-
-final class TaskOps[R, A](val e: Eff[R, A]) extends AnyVal {
-
-  def runTaskMemo[U](cache: Cache)(implicit m: Member.Aux[Memoized, R, U], task: TimedTask |= U): Eff[U, A] =
-    TaskEffect.runTaskMemo(cache)(e)
-
-  def taskAttempt(implicit task: TimedTask /= R): Eff[R, Throwable Either A] =
-    TaskInterpretation.taskAttempt(e)
-
-  def taskMemo(key: AnyRef, cache: Cache)(implicit task: TimedTask /= R): Eff[R, A] =
-    TaskInterpretation.taskMemo(key, cache, e)
-
-  def runAsync(implicit sexs: ScheduledExecutorService, ev: Eff[R, A] =:= Eff[Fx.fx1[TimedTask], A]): Task[A] =
-    TaskInterpretation.runAsync(e)
-
-  def runSequential(implicit sexs: ScheduledExecutorService, ev: Eff[R, A] =:= Eff[Fx.fx1[TimedTask], A]): Task[A] =
-    TaskInterpretation.runSequential(e)
 }
 
 object TaskInterpretation extends TaskInterpretation
