--- monix/shared/src/main/scala/org/atnos/eff/syntax/addon/monix/package.scala
@@ -1,5 +1,0 @@
-package org.atnos.eff.syntax.addon
-
-import org.atnos.eff.addon.monix.TaskEffect
-
-package object monix extends TaskEffect
