+++ scalaz/src/main/scala/org/atnos/eff/addon/scalaz/concurrent/TaskEffect.scala
@@ -147,8 +147,6 @@ trait TaskInterpretation extends TaskTypes {
       }
     )
 
-  implicit final def toTaskOps[R, A](e: Eff[R, A]): TaskOps[R, A] = new TaskOps[R, A](e)
-
   def runTaskMemo[R, U, A](cache: Cache)(effect: Eff[R, A])(implicit m: Member.Aux[Memoized, R, U], async: TimedTask |= U): Eff[U, A] = {
     interpret.translate(effect)(new Translate[Memoized, U] {
       def apply[X](mx: Memoized[X]): Eff[U, X] =
@@ -159,24 +157,6 @@ trait TaskInterpretation extends TaskTypes {
     })
   }
 
-}
-
-final class TaskOps[R, A](val e: Eff[R, A]) extends AnyVal {
-
-  def runTaskMemo[U](cache: Cache)(implicit m: Member.Aux[Memoized, R, U], task: TimedTask |= U): Eff[U, A] =
-    TaskEffect.runTaskMemo(cache)(e)
-
-  def taskAttempt(implicit task: TimedTask /= R): Eff[R, Throwable Either A] =
-    TaskInterpretation.taskAttempt(e)
-
-  def taskMemo(key: AnyRef, cache: Cache)(implicit async: TimedTask /= R): Eff[R, A] =
-    TaskInterpretation.taskMemo(key, cache, e)
-
-  def runAsync(implicit sexs: ScheduledExecutorService, ec: ExecutionContext, ev: Eff[R, A] =:= Eff[Fx.fx1[TimedTask], A]): Task[A] =
-    TaskInterpretation.runAsync(e)
-
-  def runSequential(implicit sexs: ScheduledExecutorService, ec: ExecutionContext, ev: Eff[R, A] =:= Eff[Fx.fx1[TimedTask], A]): Task[A] =
-    TaskInterpretation.runSequential(e)
 }
 
 object TaskInterpretation extends TaskInterpretation
