+++ shared/src/main/scala/org/atnos/eff/package.scala
@@ -18,8 +18,6 @@ package object eff {
   object choose     extends ChooseCreation         with ChooseInterpretation
   object safe       extends SafeCreation           with SafeInterpretation
   object future     extends FutureCreation         with FutureInterpretation
-  @deprecated("The Async effect will be removed in favor of concrete asynchronous effects, like TimedFuture.", since = "2.3.0")
-  object async      extends AsyncCreation          with AsyncInterpretation
   object memo       extends MemoCreation           with MemoInterpretation
   object batch      extends Batch
