+++ twitter/src/main/scala/org/atnos/eff/addon/twitter/TwitterFutureEffect.scala
@@ -127,8 +127,6 @@ trait TwitterFutureInterpretation extends TwitterFutureTypes {
       prom
     }
 
-  implicit final def toTwitterFutureOps[R, A](e: Eff[R, A]): TwitterFutureOps[R, A] = new TwitterFutureOps[R, A](e)
-
   /**
     * Memoize future values using a cache
     *
@@ -159,25 +157,6 @@ trait TwitterFutureInterpretation extends TwitterFutureTypes {
     })
   }
 
-}
-
-final class TwitterFutureOps[R, A](val e: Eff[R, A]) extends AnyVal {
-
-  def runTwitterFutureMemo[U](cache: Cache)(implicit memMember: Member.Aux[Memoized, R, U],
-                                            futMember: TwitterTimedFuture |= U): Eff[U, A] =
-    TwitterFutureEffect.runFutureMemo(cache)(e)(memMember, futMember)
-
-  def twitterFutureAttempt(implicit future: TwitterTimedFuture /= R): Eff[R, Throwable Either A] =
-    TwitterFutureInterpretation.futureAttempt(e)
-
-  def twitterFutureMemo(key: AnyRef, cache: Cache)(implicit future: TwitterTimedFuture /= R): Eff[R, A] =
-    TwitterFutureInterpretation.futureMemo(key, cache, e)
-
-  def runAsync(implicit pool: FuturePool, sexs: ScheduledExecutorService, ev: Eff[R, A] =:= Eff[Fx.fx1[TwitterTimedFuture], A]): Future[A] =
-    TwitterFutureInterpretation.runAsync(e)
-
-  def runSequential(implicit pool: FuturePool, sexs: ScheduledExecutorService, ev: Eff[R, A] =:= Eff[Fx.fx1[TwitterTimedFuture], A]): Future[A] =
-    TwitterFutureInterpretation.runSequential(e)
 }
 
 object TwitterFutureInterpretation extends TwitterFutureInterpretation
